Improve archive-wal robustness and log messages
Display clearer error messages when discarding a WAL file.
Also, WAL files that are older than the first backup are
now moved in the 'errors' directory: FAILED backups are now
ignored in this process.

TODO: Add a method to check the presence of files in the error
directory.

Fixes: #24

+++ barman/wal_archiver.py
@@ -120,31 +120,26 @@ class WalArchiver(RemoteStatusMixin):
                             log=False)
             found = True
 
-            # If there are no available backups ...
-            if first_backup is None:
-                # ... delete xlog segments only for exclusive backups
-                if BackupOptions.CONCURRENT_BACKUP \
-                        not in self.config.backup_options:
-                    # Skipping history files
-                    if not xlog.is_history_file(wal_info.name):
-                        output.info("\tNo base backup available."
-                                    " Trashing file %s"
-                                    " from server %s",
-                                    wal_info.name, self.config.name)
-                        os.unlink(wal_info.orig_filename)
-                        continue
-            # ... otherwise
-            else:
-                # ... delete xlog segments older than the first backup
-                if wal_info.name < first_backup.begin_wal:
-                    # Skipping history files
-                    if not xlog.is_history_file(wal_info.name):
-                        output.info("\tOlder than first backup."
-                                    " Trashing file %s"
-                                    " from server %s",
-                                    wal_info.name, self.config.name)
-                        os.unlink(wal_info.orig_filename)
-                        continue
+            # Delete the xlog segment if no backup is present and
+            # backup strategy is not concurrent and
+            # the wal file is not a history file
+            if (first_backup is None and
+                    BackupOptions.CONCURRENT_BACKUP not in
+                    self.config.backup_options and
+                    not xlog.is_history_file(wal_info.name)):
+                output.info("\tNo base backup available. "
+                            "Trashing file %s from server %s",
+                            wal_info.name, self.config.name)
+                os.unlink(wal_info.orig_filename)
+                continue
+            # ... otherwise move the wal file in the error directory
+            # if not relevant according to the first backup present
+            elif not self.is_wal_relevant(wal_info, first_backup):
+                error_dst = os.path.join(
+                    self.config.errors_directory,
+                    "%s.%s.error" % (wal_info.name, stamp))
+                shutil.move(wal_info.orig_filename, error_dst)
+                continue
 
             # Report to the user the WAL file we are archiving
             output.info("\t%s", wal_info.name, log=False)
@@ -199,8 +194,7 @@ class WalArchiver(RemoteStatusMixin):
 
     def archive_wal(self, compressor, wal_info):
         """
-        Archive a WAL segment from the incoming directory.
-        This function returns a WalFileInfo object.
+        Archive a WAL segment and return the updated WalFileInfo object
 
         :param compressor: the compressor for the file (if any)
         :param wal_info: WalFileInfo of the WAL file is being processed
@@ -318,6 +312,40 @@ class WalArchiver(RemoteStatusMixin):
             script = HookScriptRunner(self, 'archive_script', 'post', error)
             script.env_from_wal_info(wal_info, dst_file)
             script.run()
+
+    def is_wal_relevant(self, wal_info, first_backup):
+        """
+        Check the relevance of a WAL file according to a provided BackupInfo
+        (usually the oldest on the server) to ensure that the WAL is newer than
+        the start_wal of the backup.
+
+        :param WalFileInfo wal_info: the WAL file we are checking
+        :param BackupInfo first_backup: the backup used for the checks
+            (usually the oldest available on the server)
+        """
+
+        # Skip history files
+        if xlog.is_history_file(wal_info.name):
+            return True
+
+        # If the WAL file has a timeline smaller than the one of
+        # the oldest backup it cannot be used in any way.
+        wal_timeline = xlog.decode_segment_name(wal_info.name)[0]
+        if wal_timeline < first_backup.timeline:
+            output.info("\tThe timeline of the WAL file %s (%s), is lower "
+                        "than the one of the oldest backup of "
+                        "server %s (%s). Moving the WAL in "
+                        "the error directory",
+                        wal_info.name, wal_timeline, self.config.name,
+                        first_backup.timeline)
+            return False
+        # Manage xlog segments older than the first backup
+        if wal_info.name < first_backup.begin_wal:
+            output.info("\tOlder than first backup of server %s. "
+                        "Moving the WAL file %s in the error directory",
+                        self.config.name, wal_info.name)
+            return False
+        return True
 
     @abstractmethod
     def get_next_batch(self):
